Move file providers out from QuestLine
